refactor(policies): rename parametrised policies submodule
